+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/frontend/FrontendService.java
@@ -13,6 +13,7 @@ import de.mephisto.vpin.restclient.preferences.UISettings;
 import de.mephisto.vpin.restclient.vpx.TableInfo;
 import de.mephisto.vpin.server.emulators.EmulatorService;
 import de.mephisto.vpin.server.games.*;
+import de.mephisto.vpin.server.mame.MameService;
 import de.mephisto.vpin.server.playlists.Playlist;
 import de.mephisto.vpin.server.preferences.PreferenceChangedListener;
 import de.mephisto.vpin.server.preferences.PreferencesService;
@@ -55,6 +56,9 @@ public class FrontendService implements InitializingBean, PreferenceChangedListe
 
   @Autowired
   private EmulatorService emulatorService;
+
+  @Autowired
+  private MameService mameService;
 
   @Autowired
   private Map<String, FrontendConnector> frontendsMap; // autowiring of Frontends
@@ -433,6 +437,15 @@ public class FrontendService implements InitializingBean, PreferenceChangedListe
     tableDetails.setGameFileName(gameFileName);
     tableDetails.setGameDisplayName(gameDisplayName);
     tableDetails.setDateModified(new Date(file.lastModified()));
+
+    if (gameEmulator.isMameEmulator()) {
+      String fullName = mameService.resolveMAMENameFor(baseName);
+      if(!StringUtils.isEmpty(fullName)) {
+        tableDetails.setGameDisplayName(fullName);
+      }
+      tableDetails.setRomName(baseName);
+    }
+
     return importGame(tableDetails);
   }