+++ RELEASE_NOTES.md
@@ -1,6 +1,5 @@
 ## Release Notes 4.8.0
 
-- **WOVP Pause Menu Item**: Screenshots for portrait mode screens are not rotated anymore.
 - **Filter Settings**: The table filter panel allows you to filter by issue type now.
 - **Backglass Manager**:
   - Added support for Zen Studio tables (you must configure DOFLinx first!).
@@ -8,4 +7,13 @@
   - Added emulator name column (which also better explains why for duplicated emulators, backglasses are shown multiple times).
   - Enabled "Backglass" section for the table overview and Zen Studio tables.
 - **MAME Support**:
-  - Added import support for MAME roms. These 
+  - Added import support for MAME games/roms. 
+  - Added deletion support for MAME games.
+  - Fixed several issues for the overview when MAME emulator was selected.
+  - Added ROM, Playlist and date columns for the overview and MAME emulators.
+
+
+### Bugfixes
+
+- **WOVP Pause Menu Item**: Screenshots for portrait mode screens are not rotated anymore.
+- **Table Overview**: Fixed backup button being visible for all emulator types.